sending latest time series metrics only
* sending only latest time series metrics

* getting the latest time_series metric value and timestamp only

* changed int types from float to int

* moved datapoints creation logic to functions

+++ internal/monitors/conviva/metricresponse.go
@@ -5,7 +5,7 @@ type metricResponse struct {
 	FilterIDValuesMap map[string][]float64 `json:"filters,omitempty"`
 	Meta              *meta                `json:"meta"`
 	Tables            map[string]table     `json:"tables,omitempty"`
-	Timestamps        []float64            `json:"timestamps,omitempty"`
+	Timestamps        []int64              `json:"timestamps,omitempty"`
 	Xvalues           []string             `json:"xvalues,omitempty"`
 }
 
@@ -15,8 +15,8 @@ type table struct {
 }
 
 type meta struct {
-	Status                float64   `json:"status,omitempty"`
-	FiltersWarmup         []float64 `json:"filters_warmup,omitempty"`
-	FiltersNotExist       []float64 `json:"filters_not_exist,omitempty"`
-	FiltersIncompleteData []float64 `json:"filters_incomplete_data,omitempty"`
+	Status                int64   `json:"status,omitempty"`
+	FiltersWarmup         []int64 `json:"filters_warmup,omitempty"`
+	FiltersNotExist       []int64 `json:"filters_not_exist,omitempty"`
+	FiltersIncompleteData []int64 `json:"filters_incomplete_data,omitempty"`
 }
